+++ uk/co/petertribble/jproc/api/JProcLwpStatus.java
@@ -81,7 +81,7 @@ public class JProcLwpStatus {
      * @return the user time of this lwp
      */
     public double getutime() {
-	return (double) utime + ((double) nutime) / 1000000000.0;
+	return utime + nutime / 1000000000.0;
     }
 
     /**
@@ -91,7 +91,7 @@ public class JProcLwpStatus {
      * @return the system time of this lwp
      */
     public double getstime() {
-	return (double) stime + ((double) nstime) / 1000000000.0;
+	return stime + nstime / 1000000000.0;
     }
 
     /**
