+++ uk/co/petertribble/jproc/api/JProcStatus.java
@@ -82,7 +82,7 @@ public class JProcStatus {
      * @return the user time consumed by this process
      */
     public double getutime() {
-	return (double) utime + ((double) nutime) / 1000000000.0;
+	return utime + nutime / 1000000000.0;
     }
 
     /**
@@ -92,7 +92,7 @@ public class JProcStatus {
      * @return the system time consumed by this process
      */
     public double getstime() {
-	return (double) stime + ((double) nstime) / 1000000000.0;
+	return stime + nstime / 1000000000.0;
     }
 
     /**
@@ -102,7 +102,7 @@ public class JProcStatus {
      * @return the user time of this processes children
      */
     public double getcutime() {
-	return (double) cutime + ((double) ncutime) / 1000000000.0;
+	return cutime + ncutime / 1000000000.0;
     }
 
     /**
@@ -112,7 +112,7 @@ public class JProcStatus {
      * @return the system time of this processes children
      */
     public double getcstime() {
-	return (double) cstime + ((double) ncstime) / 1000000000.0;
+	return cstime + ncstime / 1000000000.0;
     }
 
     /**
